Add pagination to events listing.
The Eventbrite API returns events in chunks of 50, with pagination information. This pagination info is used to set WordPress pagination, in a way that only the chunk of 50 that applies to the page will be called. Currently, math is kept to a minimum by working with a fixed 10 events per page (meaning we're not supporting the WP_Query `posts_per_page` argument, nor `Settings > Reading > Blog pages show at most`). This will come in the future.