Automatically load template files in a theme's "eventbrite" folder, if present. Fixes #21.
If theme support is declared, but arguments are missing for a certain template, the plugin will look for that template (`eventbrite-single.php` or `eventbrite-index.php`) in an `eventbrite` folder in the theme's root. This allows theme devs using the default template locations to just use `add_theme_support( 'eventbrite' );` when declaring support. If arguments are passed for template locations, those will be used. Note that fatals can occur if a template location is defined that doesn't exist.