If an organizer ID is declared in Eventbrite_Query args, use that ID in the API call rather than the user ID.
This corrects issues when the organizer events are not owned by the user (the one who established the Eventbrite connection for the plugin). When `display_private` is `true`, this arg is ignored for the API call (since it has to be on the `owned_events` endpoint), and the organizer ID will be used in post-API processing for filtering.